requirements.txt
pynput>=1.7.6
pyaudio>=0.2.13
faster-whisper>=1.0.0
PyYAML>=6.0
numpy>=1.24.0
